Reorganized classes that were in root package
