+++ QKSMS/src/main/java/com/android/mms/transaction/RetrieveTransaction.java
@@ -36,7 +36,7 @@ import com.google.android.mms.pdu_alt.PduHeaders;
 import com.google.android.mms.pdu_alt.PduParser;
 import com.google.android.mms.pdu_alt.PduPersister;
 import com.google.android.mms.pdu_alt.RetrieveConf;
-import com.moez.QKSMS.MmsConfig;
+import com.moez.QKSMS.common.MmsConfig;
 import com.moez.QKSMS.mmssms.Utils;
 
 import java.io.IOException;
