+++ QKSMS/src/main/java/com/moez/QKSMS/common/google/PduLoaderManager.java
@@ -26,7 +26,7 @@ import com.google.android.mms.pdu_alt.MultimediaMessagePdu;
 import com.google.android.mms.pdu_alt.PduPersister;
 import com.google.android.mms.util_alt.PduCache;
 import com.google.android.mms.util_alt.PduCacheEntry;
-import com.moez.QKSMS.LogTag;
+import com.moez.QKSMS.common.LogTag;
 import com.moez.QKSMS.model.SlideshowModel;
 
 import java.util.Set;
