+++ QKSMS/src/main/java/com/moez/QKSMS/common/utils/MessageUtils.java
@@ -61,11 +61,11 @@ import com.google.android.mms.pdu_alt.PduPart;
 import com.google.android.mms.pdu_alt.PduPersister;
 import com.google.android.mms.pdu_alt.RetrieveConf;
 import com.google.android.mms.pdu_alt.SendReq;
-import com.moez.QKSMS.LogTag;
-import com.moez.QKSMS.MmsConfig;
+import com.moez.QKSMS.common.LogTag;
+import com.moez.QKSMS.common.MmsConfig;
 import com.moez.QKSMS.QKSMSApp;
 import com.moez.QKSMS.R;
-import com.moez.QKSMS.TempFileProvider;
+import com.moez.QKSMS.common.TempFileProvider;
 import com.moez.QKSMS.common.google.ThumbnailManager;
 import com.moez.QKSMS.common.google.UriImage;
 import com.moez.QKSMS.data.Contact;
