+++ QKSMS/src/main/java/com/moez/QKSMS/data/Contact.java
@@ -21,7 +21,7 @@ import android.provider.ContactsContract.Profile;
 import android.text.TextUtils;
 import android.util.Log;
 
-import com.moez.QKSMS.LogTag;
+import com.moez.QKSMS.common.LogTag;
 import com.moez.QKSMS.QKSMSApp;
 import com.moez.QKSMS.R;
 import com.moez.QKSMS.common.utils.PhoneNumberUtils;
