+++ QKSMS/src/main/java/com/moez/QKSMS/data/Conversation.java
@@ -24,7 +24,7 @@ import com.android.mms.transaction.MmsMessageSender;
 import com.google.android.mms.pdu_alt.PduHeaders;
 import com.moez.QKSMS.QKSMSAppBase;
 import com.moez.QKSMS.mmssms.Utils;
-import com.moez.QKSMS.LogTag;
+import com.moez.QKSMS.common.LogTag;
 import com.moez.QKSMS.QKSMSApp;
 import com.moez.QKSMS.R;
 import com.moez.QKSMS.common.google.DraftCache;
