+++ QKSMS/src/main/java/com/moez/QKSMS/exception/ContentRestrictionException.java
@@ -15,19 +15,23 @@
  * limitations under the License.
  */
 
-package com.moez.QKSMS;
+package com.moez.QKSMS.exception;
 
 /**
- * An exception that is thrown when content type is not supported.
+ * A generic exception that is thrown by checking content restriction.
  */
-public final class UnsupportContentTypeException extends ContentRestrictionException {
-    private static final long serialVersionUID = 2684128059358484321L;
+public class ContentRestrictionException extends RuntimeException {
+    private static final long serialVersionUID = 516136015813043499L;
 
-    public UnsupportContentTypeException() {
+    public ContentRestrictionException() {
         super();
     }
 
-    public UnsupportContentTypeException(String msg) {
+    public ContentRestrictionException(String msg) {
         super(msg);
+    }
+
+    public ContentRestrictionException(Exception cause) {
+        super(cause);
     }
 }
