+++ QKSMS/src/main/java/com/moez/QKSMS/model/ImageModel.java
@@ -29,11 +29,11 @@ import com.google.android.mms.MmsException;
 import com.google.android.mms.pdu_alt.PduPart;
 import com.google.android.mms.pdu_alt.PduPersister;
 import com.google.android.mms.smil.SmilHelper;
-import com.moez.QKSMS.ContentRestrictionException;
-import com.moez.QKSMS.ExceedMessageSizeException;
-import com.moez.QKSMS.LogTag;
+import com.moez.QKSMS.exception.ContentRestrictionException;
+import com.moez.QKSMS.exception.ExceedMessageSizeException;
+import com.moez.QKSMS.common.LogTag;
 import com.moez.QKSMS.QKSMSApp;
-import com.moez.QKSMS.MmsConfig;
+import com.moez.QKSMS.common.MmsConfig;
 import com.moez.QKSMS.common.google.ItemLoadedCallback;
 import com.moez.QKSMS.common.google.ItemLoadedFuture;
 import com.moez.QKSMS.common.google.ThumbnailManager;
