+++ QKSMS/src/main/java/com/moez/QKSMS/model/MediaModel.java
@@ -24,8 +24,8 @@ import android.net.Uri;
 import android.util.Log;
 import com.google.android.mms.MmsException;
 import com.google.android.mms.smil.SmilHelper;
-import com.moez.QKSMS.LogTag;
-import com.moez.QKSMS.MmsConfig;
+import com.moez.QKSMS.common.LogTag;
+import com.moez.QKSMS.common.MmsConfig;
 import org.w3c.dom.events.EventListener;
 
 import java.io.FileInputStream;
