+++ QKSMS/src/main/java/com/moez/QKSMS/model/MediaModelFactory.java
@@ -25,8 +25,7 @@ import com.google.android.mms.ContentType;
 import com.google.android.mms.MmsException;
 import com.google.android.mms.pdu_alt.PduBody;
 import com.google.android.mms.pdu_alt.PduPart;
-import com.moez.QKSMS.LogTag;
-import com.moez.QKSMS.MmsConfig;
+import com.moez.QKSMS.common.MmsConfig;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
